+++ src/components/domain/mypage/index.ts
@@ -1,0 +1,1 @@
+export * from './information'
