+++ src/components/domain/mypage/information/index.ts
@@ -1,0 +1,1 @@
+export { default as Information } from './Information'
